+++ client/src/pages/SignUp.js
@@ -1,12 +1,11 @@
 import React from "react";
 import { Component } from "react";
-// import API from "../utils/API";
 import { Col, Row, Container } from "../components/Grid";
-import Navbar from "../components/Navbar";
-// import FlightsCard from "../components/FlightsCard";
-import LoginSignForm from "./containers/LoginSignForm";
+import SiteNav from "../components/SiteNav";
+import FlightsCard from "../components/FlightsCard";
+import SignUpForm from "../components/SignUpForm";
 
-class Login extends Component {
+class SignUp extends Component {
 
     render() {
         return (
@@ -14,12 +13,12 @@ class Login extends Component {
                 <Container>
                     <Row>
                         <Col size="md-12">
-                            <Navbar />
+                            <SiteNav />
                         </Col>
                     </Row>
                     <Row>
                         <Col size="md-12">
-                            <LoginSignForm />
+                            <SignUpForm />
                         </Col>
                     </Row>
                 </Container >
@@ -27,5 +26,4 @@ class Login extends Component {
         );
     }
 }
-
-export default Login;
+export default SignUp;
